Testsuite: Turn off vect-epilogue-nomask for slp-rect-3
Without epiloque no mask it would only try HI modes, but thanks to the
epiloques nomask It tries QI mode as well which succeeds.  The xfail
then generates an xpass since the condition on it checks for HI to SI
and not QI.

So I disabled the epiloque mask since it seems to violate the conditions
the test actually wanted to test for.

gcc/testsuite/ChangeLog:

	* gcc.dg/vect/slp-reduc-3.c: Turn off epilogue-nomask.



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@278834 138bc75d-0d04-0410-961f-82ee72b054a4