Make vectorizable_operation punt early on codes it doesn't handle
vectorizable_operation returned false for codes that are handled by
vectorizable_shift, but only after it had already done a lot of work.
Checking earlier should be more efficient and avoid polluting the logs
with duplicate info.

Also, there was no such early-out for comparisons or COND_EXPRs.
Fixing that avoids a false scan-tree-dump hit with a later patch.

2019-11-29  Richard Sandiford  <richard.sandiford@arm.com>

gcc/
	* tree-vect-stmts.c (vectorizable_operation): Punt early
	on codes that are handled elsewhere.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@278848 138bc75d-0d04-0410-961f-82ee72b054a4