driver: Do not warn about ineffective `-x' option if no inputs were given
Fix an issue with the GCC driver and the `-x' option where a warning is
issued in an invocation like:

$ riscv64-linux-gnu-gcc -print-multi-directory -x c++
riscv64-linux-gnu-gcc: warning: '-x c++' after last input file has no effect
lib64/lp64d
$ 

where no inputs were given and hence the use of `-x' is irrelevant.
The statement printed is also untrue as the `-x' does not come after the
last input file given that none was given.  Do not print it then if no
inputs were supplied.

	* gcc.c (process_command): Only warn about an ineffective `-x'
	option if any input files have actually been supplied.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@278872 138bc75d-0d04-0410-961f-82ee72b054a4