Make dwarf2out punt for MODE_VECTOR_BOOL
The dwarf2 handling of vector constants currently divides the vector
into a length (number of elements) and byte element size.  This doesn't
work well for MODE_VECTOR_BOOL, where several elements are packed into
the same byte.

We should probably add a way of encoding this in future, but for now
the safest thing is to punt, like we already do for variable-length
vectors.

2019-12-10  Richard Sandiford  <richard.sandiford@arm.com>

gcc/
	* dwarf2out.c (loc_descriptor): Punt for MODE_VECTOR_BOOL.
	(add_const_value_attribute): Likewise.

gcc/testsuite/
	* gcc.target/aarch64/sve/acle/general/debug_4.c: New test.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@279165 138bc75d-0d04-0410-961f-82ee72b054a4