[Committed, testsuite] Fix PR92870
With my recent commit, I added a test that is not passing on all targets.
My change was valid for targets that have a vector/scalar shift/rotate optabs
(optab that supports vector shifted by scalar).

Since it does not seem to be easy to find out which targets would support it,
I am limiting the test to the targets that I know pass.

gcc/testsuite/ChangeLog

2019-12-12  Sudakshina Das  <sudi.das@arm.com>

	PR testsuite/92870
	* gcc.dg/vect/vect-shift-5.c: Add target to scan-tree-dump.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@279310 138bc75d-0d04-0410-961f-82ee72b054a4