PR c++/12333 - X::~X() with implicit this->.
this->X::~X() is handled by finish_class_member_access_expr and its
lookup_destructor subroutine; let's use it in cp_parser_lookup_name for the
case where this-> is implicit.

I tried replacing the other destructor code here with just the call to
lookup_destructor, but that regressed handling of naming the destructor
outside a non-static member function.

	* parser.c (cp_parser_lookup_name): Use lookup_destructor.
	* typeck.c (lookup_destructor): No longer static.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@279522 138bc75d-0d04-0410-961f-82ee72b054a4